+++ themes/theme-default/src/client/composables/index.ts
@@ -6,6 +6,7 @@ export * from './useNavigate.js'
 export * from './useNavbarConfig.js'
 export * from './useNavbarRepo.js'
 export * from './useNavbarSelectLanguage.js'
+export * from './useRelatedLinks.js'
 export * from './useScrollPromise.js'
 export * from './useSidebarItems.js'
 export * from './useThemeData.js'
